First pass at moving views.
