+++ backend/audit/views/submission_progress_view.py
